LLM-Network-Analyzer/requirements.txt
pandas
scapy
python-dotenv
openai
networkx
matplotlib
numpy
pyvis
streamlit